requirements.txt
aiofiles==23.2.1 ; python_version >= "3.11" and python_version < "4.0"
aiogram==3.5.0 ; python_version >= "3.11" and python_version < "4.0"
aiohttp==3.9.5 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.3.0 ; python_version >= "3.11" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.11" and python_version < "4.0"
beautifulsoup4==4.12.3 ; python_version >= "3.11" and python_version < "4.0"
brotli==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
cachetools==5.3.3 ; python_version >= "3.11" and python_version < "4.0"
cairocffi==1.6.1 ; python_version >= "3.11" and python_version < "4.0"
cairosvg==2.7.1 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
cohere==5.3.4 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
cssselect2==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
curl-cffi==0.6.3 ; python_version >= "3.11" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.11" and python_version < "4.0"
dependency-injector==4.41.0 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.6.1 ; python_version >= "3.11" and python_version < "4.0"
duckduckgo-search==5.3.0 ; python_version >= "3.11" and python_version < "4.0"
fastavro==1.9.4 ; python_version >= "3.11" and python_version < "4.0"
filelock==3.14.0 ; python_version >= "3.11" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.11" and python_version < "4.0"
fsspec==2024.3.1 ; python_version >= "3.11" and python_version < "4.0"
g4f[image]==0.3.0.6 ; python_version >= "3.11" and python_version < "4.0"
google-ai-generativelanguage==0.6.2 ; python_version >= "3.11" and python_version < "4.0"
google-api-core==2.18.0 ; python_version >= "3.11" and python_version < "4.0"
google-api-core[grpc]==2.18.0 ; python_version >= "3.11" and python_version < "4.0"
google-api-python-client==2.127.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.29.0 ; python_version >= "3.11" and python_version < "4.0"
google-generativeai==0.5.2 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.63.0 ; python_version >= "3.11" and python_version < "4.0"
grpcio-status==1.62.2 ; python_version >= "3.11" and python_version < "4.0"
grpcio==1.62.2 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==22.0.0 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.11" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.11" and python_version < "4.0"
httpx-sse==0.4.0 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.11" and python_version < "4.0"
huggingface-hub==0.23.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.7 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.3 ; python_version >= "3.11" and python_version < "4.0"
magic-filter==1.0.12 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.11" and python_version < "4.0"
md2tgmd @ git+https://github.com/yym68686/md2tgmd.git@e07840412a8333b44f30ff4194e3d3d64794660b ; python_version >= "3.11" and python_version < "4.0"
motor==3.4.0 ; python_version >= "3.11" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.11" and python_version < "4.0"
nodeenv==1.8.0 ; python_version >= "3.11" and python_version < "4.0"
orjson==3.10.1 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.11" and python_version < "4.0"
pillow==10.3.0 ; python_version >= "3.11" and python_version < "4.0"
prisma==0.13.1 ; python_version >= "3.11" and python_version < "4.0"
proto-plus==1.23.0 ; python_version >= "3.11" and python_version < "4.0"
protobuf==4.25.3 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.4.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.11" and python_version < "4.0"
pycryptodome==3.20.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.18.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.7.1 ; python_version >= "3.11" and python_version < "4.0"
pymongo==4.7.1 ; python_version >= "3.11" and python_version < "4.0"
pyparsing==3.1.2 ; python_version >= "3.11" and python_version < "4.0"
pypdfium2==4.29.0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
sentry-sdk==1.45.0 ; python_version >= "3.11" and python_version < "4.0"
setuptools==69.5.1 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
soupsieve==2.5 ; python_version >= "3.11" and python_version < "4.0"
tinycss2==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
tokenizers==0.19.1 ; python_version >= "3.11" and python_version < "4.0"
tomlkit==0.12.4 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.66.2 ; python_version >= "3.11" and python_version < "4.0"
types-requests==2.31.0.20240406 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.11.0 ; python_version >= "3.11" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.11" and python_version < "4.0"
webencodings==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
yarl==1.9.4 ; python_version >= "3.11" and python_version < "4.0"
